user api developed

+++ ToDoList/ToDoList/src/main/java/com/example/ToDoList/Services/TaskServices.java
@@ -1,19 +1,21 @@
 package com.example.ToDoList.Services;
 
 import com.example.ToDoList.Repository.TaskRepo;
+import com.example.ToDoList.Repository.UserRepository;
 import com.example.ToDoList.entity.Task;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
 
 import java.util.Optional;
 
 @Service
 public class TaskServices {
     @Autowired
-    TaskRepo taskRepo;
+   private TaskRepo taskRepo;
+    @Autowired
+   private UserRepository userRepo;
 
 
     public Task addTask(Task task) {
@@ -21,28 +23,32 @@ public class TaskServices {
     }
 
     public Optional<Task> getTask(Long id) {
+
         return taskRepo.findById(id);
     }
 
-    public ResponseEntity<String> deleteTask(Long id) {
-       Optional<Task> task= taskRepo.findById(id);
-       if(task.isPresent()){
-            Task data=task.get();
-            if(data.isCompleted()){
+    public ResponseEntity<String> deleteTask(String name,Long id) {
+        if (!userRepo.existsByName(name)) {
+            return new ResponseEntity<>("user not found ", HttpStatus.NOT_FOUND);
+        }
+        Optional<Task> task = taskRepo.findById(id);
+        if (task.isPresent()) {
+            Task data = task.get();
+            if (data.isCompleted()) {
                 taskRepo.deleteById(id);
-                return ResponseEntity.ok("successfully completed ");
-            }
-            else{
+                return ResponseEntity.ok("deletion successfully completed ");
+            } else {
                 return ResponseEntity.badRequest().body(" Task is not completed ");
             }
-       }
-       else{
-           return new ResponseEntity(HttpStatus.NOT_FOUND);
-       }
+        }
+        return new ResponseEntity<>("Task Not Found",HttpStatus.NOT_FOUND);
     }
 
-    public ResponseEntity<String> updateTask(Long id) {
 
+    public ResponseEntity<String> updateTask(String name, Long id) {
+        if(!userRepo.existsByName(name)){
+            return new ResponseEntity<>("Not Found User"+name,HttpStatus.NOT_FOUND);
+        }
        Optional<Task>task=taskRepo.findById(id);
        if(task.isPresent()){
            Task data=task.get();
@@ -56,11 +62,11 @@ public class TaskServices {
              return ResponseEntity.badRequest().body("Task is already completed ");
            }
        }
-       else{
-        return ResponseEntity.badRequest().body("Task Not found");
-       }
+      return new ResponseEntity<>("Task Not Found ",HttpStatus.NOT_FOUND);
 
     }
+
+
 }
